Add static IP, disconnect WiFi, limit while loop
Using a static IP saves connection/execution time and battery use.
Not disconnecting WiFi caused connection errors the second time. Possibly not a Firebeetle, but a router issue?
While loops did not have a limit, causing the script to loop indefinitely when no WiFi or MQTT server could be found.
When GPIO26 was triggered, GPIO26 does not have to be armed. Checking the mailbox when no mail arrived would cause unnecessary wake-ups.