DHCP + retry functions + wake-up port fix
Added DHCP enable/disable option, retry functions and fix for when the board does not provide the correct port that triggered the wake-up.  When this happens the script checks if either port is still HIGH. I've been using this code for a few months now and had no issues since.